DAOs modificados para que laburen con int en las claves (como la base de datos) y añadida la primer clase ABM

+++ ProyectoSuba/src/dao/BoletoColectivoDao.java
@@ -61,7 +61,7 @@ public class BoletoColectivoDao {
 		}
 	}
 	
-	public BoletoColectivo traer(long idBoletoColectivo) throws HibernateException {
+	public BoletoColectivo traer(int idBoletoColectivo) throws HibernateException {
 		BoletoColectivo objeto = null;
 		
 		try {
